CRM-20858: Fix Failing Test on Custom Field Merge
Test for merging custom fields with is_view flag on was failing. This was
happening because the same field was being used on prior tests, setting the
is_view flag at the start of the failing test. Fixed by creating a new custom
field with the is_view flag set as 1, and using that field on the test.