CRM-17281 Pledge payments : fix bug with miscalculation when assigning pledge payments.
This PR fixes a bug where payments are miscalculated with weird results. If a payment
changes the amount of future payments in such a way as to make an irregular amount
e.g 10 is split over 3 payments, the payment amount needs to be recorded as .33 not
.333333, assuming a currency with 2 decimal places. This commit
ensures the rounding takes place & is locked in by the accompanying unit test.

Note that we discussed this & agreed to use a function as a placeholder for the number
of decimal places.