CRM-12167: Finish Implementing Visibility per Price Option
Added validations so that:
- A public price field has at least one public option
- An admin price field has all admin options

Implemented removal for Select and Checkbox field types. Also added condition
to remove Admin values only if logged user doen't have appropriate
permissions.

Implemented different permissions check to see if admin options should be
deleted or not if the price set is being used for a contribution page or an
event page.

Implemented functionality so that if a user chooses Admin visibility for the
price field, all options are forced to have Admin visibility as well.

Added default visibility as public on creation, in case it's not set, and
protected access to unexisting visibility_id array key.

Moved SQL to add new visibility field to PHP upgrader for v4.7.26.