CRM-21112: Optimize Count On Relationships Tab
Relationships tab for contact's summary was loading the table's contents with
an AJAX call. This load was counting the total number of relationships by
loading ALL relationships into an array, and then counting its size, something
that brought the system to a crawl on contacts with a lot of relationships
(>20K).

Fixed by doing the count at DB level, by passing the count flag as 1 to the
BAO.