CRM-21328: Remove '-select-' Option From Visibility
As the visibility per option field is not mandatory, the user can choose the
empty option and save the field. This caused some PHP warnings to be issued on
event registration forms that used the field with empty visibility options.

Removed the empty option from the visibility per option select.