CRM-17860, CRM-18231 - JobProcessMailingTest - Re-remove require_once
Overview
--------
Remove an unnecessary line from test suite.

Before
------
The test worked when executed as part of the `api_v3` suite, but it failed
when run individually (because the `require_once` doesn't work).  Moreover,
the tests for `org.civicrm.flexmailer` failed

After
-----
The tests work on their own.

Comment
-------

This change deals entirely with the packaging/framing of a test-case.  It
neither (a) involves the runtime behavior of the application nor (b) the
semantics/behavior of the test.  Consequently, it should be safe to merge as
long as the test-runner remains happy.