[NFC] Function extraction and renaming of variable for shorter reference.
This is a function extraction towards CRM-19273. I haven't changed the logic but it is apparent that
we are changing the contribution financial type to 'the financial type of the last line
item' - which is a little dubious although it might work in practice. We should expose the field &
let people choose it