CRM-21656: Fix Backend Membership with Priceset Applies Taxes Twice
Steps to reproduce:

- Enable Taxes, Add a tax account with a 10% tax
- Add the tax account to the Membership financial type
- Create a new priceset for memberships, with radio options for the main membership options.
- From a contact record, add a new membership for that contact, select an option from the priceset.

Notice how to total amount is OK, but when we view the contribution, the line
item shows a unit_price with the tax_amount included, so then the total_line
ends up adding the taxes again.

This was caused by a hack for entering backend memberships using the default
dropdown options (not a priceset), which makes it possible to enter an arbitrary
total_amount directly.