CRM-21648: Only Allow To Add Contacts in Configured Roles
We were allowing to add contacts in bulk in any role, even if it was not
configured as a case role in the case's case type. This was causing errors on
case detail view, on 'People' tab, when more than one contact had a role not
defined on the case-type.

Fixed by only allowing to add contacts in the case's configured roles, as
defined by the case-type.