CRM-19784 remove yahoo geocoder
This also includes some tidy up of the way in which the geocoder
is called. A specific goal was to avoid a fatal error where a geocoder
is called that no longer exists. This applies specifically to the
Yahoo class but also is presumably the reason for the hard-coded
hack (which I removed) around the non-existent open street map class.