CRM-17647 improve testing on the thousand separator.
As part of fixing thousand separator issues we can use the deprecated flag to identify when the
BAO_Contribution::create is called without skip_clean_money as we want to move all cleaning to
close to submit. First, however, we need to stop calling that function directly in the test suite
(or only call with skip_clean_money if we are specifically testing it).

Am checking how jenkins likes this format (which was no good on 5.3) & will change others if cheerful