CRM-21648: Filter Available Case Roles with Selected Contact Types
We should only allow the user to select roles that are appropriate for the
contact types allowed for the role (eg. if you select an organization you
should not be able to add it as sibling to the case).

Fixed by filtering available case roles so that if more than one contact type
is selected, only roles that allow all contact types can be selected, and if
all contacts have the same type, only roles that allow that contact type can
be selected.