(dev/core#285) Fixed second membership repeating reminder
This fixes the code for reminders that need to be sent again when the
entity used for the date comparison changes. This is just membership
right now. With this fix, when a membership is updated, the scheduled
reminder conditions all act as if no reminders have been sent yet. This
means that if you setup a schedule reminder that is checking against
the membership end date, the schedule reminder will go out at the new
membership end date even though one already went out with the previous
end date.