CRM-21707 use serialisation metadata from api basic_create_fallback.
In order to add a unit test I enabled the same in-DAO handling on the Group BAO create
(since core api don't use the fallback). It allowed me to remove cruft and
to remove comments about removing cruft.

With this in place a custom entity (in an extension) could define

    <pseudoconstant>
      <optionGroupName>group_type</optionGroupName>
    </pseudoconstant>
    <serialize>SEPARATOR_BOOKEND</serialize>

and that field would accept an array of valid names that map to group_types
and the api would handle the serialisation. api v4 supports this so good prep