CRM-21648: Implement Logic Including Undefined Contact Types
If possible contct types is set as All, API may not return the field, as it is
empty, causing it to be undefined. This in turn caused the contcact type to be
falsely skipped, as undefined is evaluated as different from an empty string.

Fixed by taking into account undefined contact types for available
relationships and treating them as if they were empty.