composer.json - De-fork dependency, marcj/topsort
When this dependency was originally added, we needed a few patches (for PHP 5.3 compatibility)
and initially used a forked version of library. Of course, it's undesirable to use a fork
in the long term (e.g. harder to apply upgrades; harder to merge into other build processes).

In the intervening period, upstream has merged the patches for PHP 5.3, and
we've politely asked downstream to get over PHP 5.3, so we're covered on
both ends.  Let's get back on the mainline branch!