CRM-21804 improve error handling on duplicate contribution.
Duplicate contributions are a really not handled well anywhere in the form layer.
The CRM_Core_Error winds up using the fatal template, which has the same end
result for form users as if an exception were thrown, let's switch.

Note that this appears to be the only scenario when an error would
be returned so we can simply drop that handling