set-version.php - Update civicrm-version.php (if applicable)
Currently, this file is produced by GenCode, so it's not critical to update it via `set-version.php`. However, it's
a bit easier to test the process if this file is kept up-to-date.

In the near future, this file may be reworked to read from the canonical
source.  However, the search/replace should still be safe then.