Forms/Sessions - Store state via Civi::cache() instead of CRM_Core_BAO_Cache
If you enable redis/memcache and use a CiviQuickForm, the state is managed
via `CRM_Core_BAO_Cache::setItem()` and `::getItem()` using multi-level
cache.  That means the state is stored twice, once in redis/memache and once
in MySQL's `civicrm_cache`. With this patch, the state is stored in
*either* redis/memcache *xor* MySQL.

The motivation is to make it easier to use Civi with a read-only MySQL
database.

Personally, the behavior of `CRM_Core_BAO_Cache` feels like pointless
complexity, so I've half a mind to just patch `getItem()`/`setItem()` to
remove the double-caching.  However, I don't want to be responsible for
immediately hunting+fixing all the edge-cases that might implicitly depend
on it.

Consequently, this PR takes the approach of moving one dataset (form/session
state) away from `CRM_Core_BAO_Cache` and toward `Civi::cache('session')`.
That cache is specified as using fallbacks (memcache or else redis or else
MySQL cache or else thread-local array).

There is a potential counter-argument that multi-level cache could be
desirable -- e.g.  try to get the greater speed of reading memcache/redis
with the greater durability of MySQL.  But doing that adds complexity and
increases the cost of writes.  Maybe it would be worthwhile for read-heavy
data...  but form-state isn't read-heavy.  It's an even 1:1 mix of
reads:writes.  (*User requests form => read the state, process a bit, update
the state.*)

I'm not opposed to the existence of multi-level caches, but they shouldn't
be hard-coded for all caching.  Instead, one might use the *Decorator*
pattern, e.g.:

```php
$multiLevelCache = new CRM_Utils_Cache_Chained([
  new CRM_Utils_Cache_Redis(...),
  new CRM_Utils_Cache_SqlGroup(...),
])
```