WIP - CRM_Core_DAO - Manage RW/RO data source
The DB_DataObject subsystem has a feature for managing multiple database
connections.  It's based on the idea that every *instance* of
`DB_DataObject` can be associated with a different DB by saying:

```php
$dao->_database = 'my-logical-db-name';
$dao->_connect();
```

This patch makes a few changes to allow DAO's to use different RW/RO
database credentials:

1.  It declares multiple allowed values of `_database` (`rw`, `ro`, `cms`).

2.  It defines a default value of `_database` for `CRM_Core_DAO` and all
subclasses.  The default is managed as a system-wide stack variable.

3.  It fixes a problem where `CRM_Core_DAO::singleValueQuery()` was only
able to query one data-source -- it used a static-cache to avoid creating
extra copies of `CRM_Core_DAO`, but this copy was locked into a single
`_database`.  We now keep multiple copies of `CRM_Core_DAO` (one for each
active `_database`).

I'm on the fence about whether it's better to use this functionality of
`DB_DataObject` or to discard it for our own connection-management.  Some
considerations:

- Some of the initialization (`sql_mode`/`utf8`/`@unique_id`) isn't applied when
  you switch a non-default DSN.

- Overall, it's confusing to trace the DB connection setup because we haven't been
  using half the options/flows permitted by `DB_DataObject`.

- It's possible to have multiple DAO instances floating around at the same time, with
  each bound to a different database. This can be good (if your logic is written
  to juggle them correctly), but it could also be bad (it's usually not desirable
  policy, and it may be quite easy to do this by accident).

- If (for some ungodly reason) a third-party extended `DB_DataObject` but
  wasn't using `CRM_Core_DAO`s connection, then they should still work.