footer.tpl - The release notes URL should be more maintainable
Overview
----------------------------------------
Change the URL formula pointing to the release notes.

Before
----------------------------------------
The formula was `https://github.com/civicrm/civicrm-core/blob/{$version}/release-notes/{$version}.md`. Because the URL
points to a tag on Github, it's hard to make any corrections if something has changed or gone wrong.

After
----------------------------------------
The formula is `https://download.civicrm.org/about/{$version}`, which corresponds to
[civicrm-dist-manager's AboutController](https://github.com/civicrm/civicrm-dist-manager/blob/master/src/CiviDistManagerBundle/Controller/AboutController.php).