CRM_Utils_Check - Generate a `notice` if MySQL doesn't support storage of accents
This is a follow-up to #12060.  I'm submitting it for discussion -- but I'm not
actually confident in the change, so please don't merge unless you affirmatively
agree.  Let me recount the context.

PR #12060 aimed to address an issue where `testInternationalStrings()` failed on
some systems (`test-ubu1604-1`) but succeeds on others (`test-ubu1204-5`).  The
content of `character_set_database` was suggested as a distinguishing
characteristic, so #12060 basically says, "Don't test if we don't expect it to work."
That's nice for developers because it reduces the noisiness of the test suite.

But it also reduces the impact of the `testInternationalStrings()` -- because a
non-working environment is considered passing.  Is it legit to say "it's not the
devteam's problem if some environments don't work"?  Yeah.  But then we should
finish kicking the can over to the sysadmin so they can fix their system.  That's
what this patch does.

Both 12060 and this patch are premised on the correctness of
`supportStorageOfAccents()`.  I'm not really confident in it, though.  On my local
system, `supportStorageOfAccents()` returns `FALSE`, but I'm able to create and view
a contact with last name `Iñtërnâtiônàlizætiøn` -- via either web UI or
 `cv api contact.create`. That casts some doubt on `supportStorageOfAccents()`,
but it's not decisive (maybe check is important for some other reason/use-case).

In short -- if `supportStorageOfAccents()` is working correctly, then we should
merge this and notify site-admins.  If `supportStorageOfAccents()` is not working
correctly, then we probably need something else to fix #12060.