Fix Relationship Type form to use the api in postProcess in preparation to support custom data.
Now that we have custom-data-on-any forms that use the api to save can potentially support custom data. This is a
preparatory tidy-up commit. It switches to the api & does minor tidy ups, including using metadata and addField to
add fields. This last part does slightly alter the wording on the form. I think that is OK?