Remove inheritance of MembershipConfig form from MembershipStatus form.
This is part of trying to move towards an entity form (ie. support custom data) for MembershipType.

Currently the 2 forms share a parent class - but really the add buttons is the only shared functionality.

This change uses the EntityTrait to add the buttons as well. There is some minor tidy up in there
but I tried to keep it very limited.

I switched only one field over to being added via addField, the parent mechanism so far