Fix display of conflicts on duplicate screen.
There are 2 bulk dedupe scenarios
1) dedupe selected
2) dedupe all

In the case of the former, but not the latter, the display to which contacts
are redirected should be filted by 'selected'

Currently the selected filter is being applied whenever conflicts is true, resulting in an
empty form in the latter case.

This fix disambiguates the 2 concepts & attempts to rationalise & clarify related parameters in the flow.