Fail more gracefully when attempting to use a soft credit filter but not including soft credits.
For better or worse this report munges soft credits & contributions in a very wierd way.

It currently hard fails if you try to search for 'contributions only' but include
a soft credit criteria, although there appears to have been an effort to block this in the past.

It might be nice to add that filter but restructuring the report to make it testable
should be a pre-requisite for that - so for now, let's address the fatal