(dev/core#183) Change temp table naming for reports
Testing notes - This is a pretty formulaic change.  The main concern is some kind of typo or bad table name causing a
hard-fail.  To control that risk, we just run the code in a facile way:

* Navigate to "Reports => Contributions => Lybunt"
* Add a "Filter" by "Group"

We do this procedure under a few circumstances:

* With the original code
* With a purposefully bad edit (provoking an expected error)
* With the new code

In this commit, we changed two lines, so we do the above once for each line.