(dev/core#174) civicrm_cache - Mandate change from DATETIME to TIMESTAMP
v4.7.20 updated the definition of `created_date` and `expired_date` so that
new installs would default to TIMESTAMP instead of DATETIME.  Status-checks
and DoctorWhen have been encouraging a transition, but it wasn't mandated,
and there was little urgency...  because `expired_date` was ignored, and
adhoc TTLs on `created_date` had generally long windows.  Now that we're
using `expired_date` in more important ways for 5.4 (dev/core#174), we want
to ensure that these values are handled precisely and consistently.

Before
------------------------------
* On new deployments, `created_date` and `expired_date` are TIMESTAMP.
* On upgraded deployments, `created_date` and `expired_date` may
  be using DATETIME or TIMESTAMP (depending on the particular history
  and administrative decisions).

After
------------------------------
* On all deployments, `created_date` and `expired_date` are TIMESTAMP.